Move functions from scripts.js to domUpdates.js
