+++ src/scripts.js
@@ -1,6 +1,7 @@
 import './css/base.scss';
 import './css/styles.scss';
 
+import domUpdates from './domUpdates';
 import recipeData from './data/recipes';
 import ingredientData from './data/ingredients';
 import users from './data/users';
@@ -19,7 +20,9 @@ let user, pantry;
 window.onload = onStartup();
 
 homeButton.addEventListener('click', cardButtonConditionals);
-favButton.addEventListener('click', viewFavorites);
+favButton.addEventListener('click', function() {
+  domUpdates.viewFavorites(user, cookbook);
+});
 cardArea.addEventListener('click', cardButtonConditionals);
 
 function onStartup() {
@@ -29,74 +32,70 @@ function onStartup() {
   });
   user = new User(userId, newUser.name, newUser.pantry)
   pantry = new Pantry(newUser.pantry)
-  populateCards(cookbook.recipes);
-  greetUser();
-}
-
-function viewFavorites() {
-  if (cardArea.classList.contains('all')) {
-    cardArea.classList.remove('all')
-  }
-  if (!user.favoriteRecipes.length) {
-    favButton.innerHTML = 'You have no favorites!';
-    populateCards(cookbook.recipes);
-    return
-  } else {
-    favButton.innerHTML = 'Refresh Favorites'
-    cardArea.innerHTML = '';
-    user.favoriteRecipes.forEach(recipe => {
-      cardArea.insertAdjacentHTML('afterbegin', `<div id='${recipe.id}'
-      class='card'>
-      <header id='${recipe.id}' class='card-header'>
-      <label for='add-button' class='hidden'>Click to add recipe</label>
-      <button id='${recipe.id}' aria-label='add-button' class='add-button card-button'>
-      <img id='${recipe.id}' class='add'
-      src='https://image.flaticon.com/icons/svg/32/32339.svg' alt='Add to
-      recipes to cook'></button>
-      <label for='favorite-button' class='hidden'>Click to favorite recipe
-      </label>
-      <button id='${recipe.id}' aria-label='favorite-button' class='favorite favorite-active card-button'>
-      </button></header>
-      <span id='${recipe.id}' class='recipe-name'>${recipe.name}</span>
-      <img id='${recipe.id}' tabindex='0' class='card-picture'
-      src='${recipe.image}' alt='Food from recipe'>
-      </div>`)
-    })
-  }
-}
-
-function greetUser() {
-  const userName = document.querySelector('.user-name');
-  userName.innerHTML =
-  user.name.split(' ')[0] + ' ' + user.name.split(' ')[1][0];
-}
-
-function favoriteCard(event) {
-  let specificRecipe = cookbook.recipes.find(recipe => {
-    if (recipe.id  === Number(event.target.id)) {
-      return recipe;
-    }
-  })
-  if (!event.target.classList.contains('favorite-active')) {
-    event.target.classList.add('favorite-active');
-    favButton.innerHTML = 'View Favorites';
-    user.addToFavorites(specificRecipe);
-  } else if (event.target.classList.contains('favorite-active')) {
-    event.target.classList.remove('favorite-active');
-    user.removeFromFavorites(specificRecipe)
-  }
+  domUpdates.populateCards(cookbook.recipes, user);
+  domUpdates.greetUser(user);
 }
 
 function cardButtonConditionals(event) {
   if (event.target.classList.contains('favorite')) {
-    favoriteCard(event);
+    domUpdates.favoriteCard(event, cookbook, user);
   } else if (event.target.classList.contains('card-picture')) {
     displayDirections(event);
   } else if (event.target.classList.contains('home')) {
     favButton.innerHTML = 'View Favorites';
-    populateCards(cookbook.recipes);
+    domUpdates.populateCards(cookbook.recipes, user);
   }
 }
+
+// function viewFavorites() {
+//   if (cardArea.classList.contains('all')) {
+//     cardArea.classList.remove('all')
+//   }
+//   if (!user.favoriteRecipes.length) {
+//     favButton.innerHTML = 'You have no favorites!';
+//     populateCards(cookbook.recipes);
+//     return
+//   } else {
+//     favButton.innerHTML = 'Refresh Favorites'
+//     cardArea.innerHTML = '';
+//     user.favoriteRecipes.forEach(recipe => {
+//       cardArea.insertAdjacentHTML('afterbegin', `<div id='${recipe.id}'
+//       class='card'>
+//       <header id='${recipe.id}' class='card-header'>
+//       <label for='add-button' class='hidden'>Click to add recipe</label>
+//       <button id='${recipe.id}' aria-label='add-button' class='add-button card-button'>
+//       <img id='${recipe.id}' class='add'
+//       src='https://image.flaticon.com/icons/svg/32/32339.svg' alt='Add to
+//       recipes to cook'></button>
+//       <label for='favorite-button' class='hidden'>Click to favorite recipe
+//       </label>
+//       <button id='${recipe.id}' aria-label='favorite-button' class='favorite favorite-active card-button'>
+//       </button></header>
+//       <span id='${recipe.id}' class='recipe-name'>${recipe.name}</span>
+//       <img id='${recipe.id}' tabindex='0' class='card-picture'
+//       src='${recipe.image}' alt='Food from recipe'>
+//       </div>`)
+//     })
+//   }
+// }
+
+// function favoriteCard(event) {
+//   let specificRecipe = cookbook.recipes.find(recipe => {
+//     if (recipe.id  === Number(event.target.id)) {
+//       return recipe;
+//     }
+//   })
+//   if (!event.target.classList.contains('favorite-active')) {
+//     event.target.classList.add('favorite-active');
+//     favButton.innerHTML = 'View Favorites';
+//     user.addToFavorites(specificRecipe);
+//   } else if (event.target.classList.contains('favorite-active')) {
+//     event.target.classList.remove('favorite-active');
+//     user.removeFromFavorites(specificRecipe)
+//   }
+// }
+
+
 
 
 function displayDirections(event) {
@@ -132,37 +131,37 @@ function displayDirections(event) {
   })
 }
 
-function getFavorites() {
-  if (user.favoriteRecipes.length) {
-    user.favoriteRecipes.forEach(recipe => {
-      document.querySelector(`.favorite${recipe.id}`).classList.add('favorite-active')
-    })
-  } else return
-}
+// function getFavorites() {
+//   if (user.favoriteRecipes.length) {
+//     user.favoriteRecipes.forEach(recipe => {
+//       document.querySelector(`.favorite${recipe.id}`).classList.add('favorite-active')
+//     })
+//   } else return
+// }
 
-function populateCards(recipes) {
-  cardArea.innerHTML = '';
-  if (cardArea.classList.contains('all')) {
-    cardArea.classList.remove('all')
-  }
-  recipes.forEach(recipe => {
-    cardArea.insertAdjacentHTML('afterbegin', `<div id='${recipe.id}'
-    class='card'>
-        <header id='${recipe.id}' class='card-header'>
-          <label for='add-button' class='hidden'>Click to add recipe</label>
-          <button id='${recipe.id}' aria-label='add-button' class='add-button card-button'>
-            <img id='${recipe.id} favorite' class='add'
-            src='https://image.flaticon.com/icons/svg/32/32339.svg' alt='Add to
-            recipes to cook'>
-          </button>
-          <label for='favorite-button' class='hidden'>Click to favorite recipe
-          </label>
-          <button id='${recipe.id}' aria-label='favorite-button' class='favorite favorite${recipe.id} card-button'></button>
-        </header>
-          <span id='${recipe.id}' class='recipe-name'>${recipe.name}</span>
-          <img id='${recipe.id}' tabindex='0' class='card-picture'
-          src='${recipe.image}' alt='click to view recipe for ${recipe.name}'>
-    </div>`)
-  })
-  getFavorites();
-};
+// function populateCards(recipes) {
+//   cardArea.innerHTML = '';
+//   if (cardArea.classList.contains('all')) {
+//     cardArea.classList.remove('all')
+//   }
+//   recipes.forEach(recipe => {
+//     cardArea.insertAdjacentHTML('afterbegin', `<div id='${recipe.id}'
+//     class='card'>
+//         <header id='${recipe.id}' class='card-header'>
+//           <label for='add-button' class='hidden'>Click to add recipe</label>
+//           <button id='${recipe.id}' aria-label='add-button' class='add-button card-button'>
+//             <img id='${recipe.id} favorite' class='add'
+//             src='https://image.flaticon.com/icons/svg/32/32339.svg' alt='Add to
+//             recipes to cook'>
+//           </button>
+//           <label for='favorite-button' class='hidden'>Click to favorite recipe
+//           </label>
+//           <button id='${recipe.id}' aria-label='favorite-button' class='favorite favorite${recipe.id} card-button'></button>
+//         </header>
+//           <span id='${recipe.id}' class='recipe-name'>${recipe.name}</span>
+//           <img id='${recipe.id}' tabindex='0' class='card-picture'
+//           src='${recipe.image}' alt='click to view recipe for ${recipe.name}'>
+//     </div>`)
+//   })
+//   getFavorites();
+// };
